server/requirements.txt
Django==3.1.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-cors-headers==3.6.0
psycopg2==2.8.6
django-extensions==3.1.0
djangorestframework-simplejwt==4.6.0
Pillow==8.1.2
py-moneyed==1.2


